backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.2.0
django-filter==23.5 
gunicorn==21.2.0
whitenoise==6.6.0
environs==9.5.0
psycopg2==2.9.9
dj-database-url==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1 
django-allauth==0.61.1
dj_rest_auth==6.0.0
Unidecode==1.3.8
unicode-slugify==0.1.5 
django-rest-swagger==2.2.0 
drf-spectacular==0.27.2 
django-shortuuidfield==0.1.3
asgiref==3.9.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-filter==23.5
django-rest-swagger==2.2.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
environs==9.5.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.26.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
setuptools==69.1.1
shortuuid==1.0.13
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
unicode-slugify==0.1.5
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.4.0
wheel==0.43.0
whitenoise==6.6.0
django-cache-url==3.4.4
django-redis==5.4.0
redis==5.0.0
django-elasticsearch-dsl==7.3 
django-elasticsearch-dsl-drf==0.22.5 
shortuuid==1.0.13 
drf-spectacular-sidecar==2025.4.1
django-debug-toolbar==5.1.0

channels==4.3.1
channels-redis==4.3.0
redis==5.0.0
django-redisboard==8.3.0
daphne==4.2.1